update charts for gtm to work in deployment
since `GOOGLE_TAG_MANAGER_ID` is being referenced through an environment
variable, it needed to be added as a deployment variable as well. we
only need this value accessible in production, so I was able to remove it
from the local env. the value should return as "undefined" and "" in the
local and staging environments, respectively.

related: https://github.com/scientist-softserv/phenovista-digital-storefront/pull/37