hopefully gtm works in prod for real now
I thought that the reason the GTM ID wasn't showing on production after
deploying my branch (https://github.com/scientist-softserv/phenovista-digital-storefront/pull/38)
to production was because the deploy changes needed to be in main to take
effect. however, after merging the above to main and deploying again, the
GTM ID still doesn't show up.

I now believe the issue is related to the `NEXT_PUBLIC` prefix. Added it
to the GTM ID so that the client should be able to pick up the variable.