wip: attempt to fix internal server error on staging
when this branch was deployed to staging, I received and internal server error
while trying to sign in. after allowing access, I was redirected back to
https://phenovista-staging.softserv.cloud/api/auth/error.

I found the initial code for `redirect` at
https://github.com/nextauthjs/next-auth/blob/main/packages/core/src/lib/init.ts#L38-L42.
I'm adding it in hopes to bypass the server error so I can continue testing on staging.