revert to select drop downs instead of radios
this is a hotfix so that we're able to view the page on production at
least.

will be looking into why the radio isn't working for Ready-2-Go
AAV Gene Delivery Assay Service. the error states "No widget 'radio' for
type 'array'", but there isn't a { type: array } in options.fields for
this service.

- ref: https://github.com/scientist-softserv/phenovista-digital-storefront/issues/51