application specific webhook names
webhooks are tied to a specific client application on the scientist.com
marketplace. currently, they're all titled "webstore", which means a user
can have multiple webhooks that are named the same, but refer to different
applications. this change makes each webhook name unique by referencing
the name of the client application.

- related: https://github.com/assaydepot/rx/issues/25317